Python_assignment02/ai-api-integration/requirements.txt
# AI API Integration - 2026 Project Requirements
google-genai>=1.2.0
cohere>=5.0.0
groq>=1.1.1
huggingface-hub>=1.7.1
python-dotenv>=1.2.2